backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
openai>=2.21.0
passlib==1.7.4
bcrypt==4.1.1
PyJWT==2.11.0
python-dotenv==1.0.0
pymssql>=2.2.8
fpdf2>=2.7.0
